Added taxes round tests and improved algorithm

+++ salestaxes-api/src/main/java/com/id/salestaxesapi/api/IPrice.java
@@ -1,6 +1,7 @@
 package com.id.salestaxesapi.api;
 
 import com.id.salestaxesapi.impl.Currency;
+import java.math.BigDecimal;
 
 /**
  * The interface for the price of the item
@@ -18,6 +19,6 @@ public interface IPrice {
      * The value
      * @return 
      */
-    public double getValue();
+    public BigDecimal getValue();
 
 }
